+++ src/be/nikiroo/utils/test_code/StringUtilsTest.java
@@ -225,8 +225,8 @@ class StringUtilsTest extends TestLauncher {
 			@Override
 			public void test() throws Exception {
 				String orig = "test";
-				String zipped = StringUtils.base64(orig, true);
-				String unzipped = StringUtils.unbase64s(zipped, true);
+				String zipped = StringUtils.zip64s(orig);
+				String unzipped = StringUtils.unzip64s(zipped);
 				assertEquals(orig, unzipped);
 			}
 		});